requirements.txt
Flask==2.3.4
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.1
gunicorn==21.2.0
pdfminer.six==2023.8.0
python-dotenv==1.0.0